@@ -13,19 +13,19 @@
     "license": "Apache-2.0",
     "authors": [
         {
-            "name": "ideaa",
-            "email": "ideaa@qq.com",
-            "homepage": "http://ideaa.sinaapp.com",
+            "name": "wonli",
+            "email": "wonli@live.com",
+            "homepage": "http://www.crossphp.com",
             "role": "Developer"
         }
     ],
     "require": {
-        "php": ">=5.2.0"
+        "php": ">=5.3.0"
     },
     "require-dev": {
         "phpunit/phpunit": "3.7.*"
     },
     "autoload": {
-        "psr-0": {"crossphp\\": "crossphp/"}
+        "psr-4": {"crossphp\\": "crossphp/"}
     }
 }
