@@ -23,8 +23,5 @@ defined('PROJECT_PATH') or die("undefined PROJECT_PATH");
  */
 define('APP_PATH_DIR', rtrim(PROJECT_PATH, DS).DS.'app');
 
-/**
- * require cross && loader
- */
-require CP_PATH .'core/Cross.php';
 require CP_PATH .'core/Loader.php';
+cross\core\Loader::init();
