@@ -3,29 +3,37 @@
  * @Auth <wonli@live.com>
  * Class Cache
  */
+namespace cross\cache;
+
+use cross\exception\CoreException;
+
 class CoreCache
 {
-    static $cache_type = array(1=>'file', 2=>'memcache', 3=>'redis');
+    /**
+     * 缓存类型
+     *
+     * @var array
+     */
+    static $cache_type = array(1 => 'file', 2 => 'memcache', 3 => 'redis');
 
     /**
      * 实例化缓存类
      *
      * @param $cache_config
-     * @return FileCache|Memcache|RedisCache
+     * @return FileCache|MemcacheBase|RedisCache
      * @throws CoreException
      */
     static function factory($cache_config)
     {
-        switch($cache_config["type"])
-        {
+        switch ($cache_config["type"]) {
             case 1:
-                return new FileCache( $cache_config );
+                return new FileCache($cache_config);
 
             case 2:
-                return new MemcacheBase( $cache_config );
+                return new MemcacheBase($cache_config);
 
             case 3:
-                return new RedisCache( $cache_config );
+                return new RedisCache($cache_config);
 
             default :
                 throw new CoreException("不支持的缓存");
