@@ -3,6 +3,11 @@
  * @Auth wonli <wonli@live.com>
  * Class FileCache
  */
+namespace cross\cache;
+
+use cross\core\Helper;
+use cross\i\CacheInterface;
+
 class FileCache implements CacheInterface
 {
     /**
@@ -21,9 +26,9 @@ class FileCache implements CacheInterface
 
     function __construct($cache_config)
     {
-        $file_ext = isset($cache_config['file_ext'])?$cache_config['file_ext']:'.html';
-        $this->cache_file  = $cache_config['cache_path'].DS.$cache_config['key'].$file_ext;
-        $this->expire_time = isset($cache_config ['expire_time'])?$cache_config ['expire_time']:3600;
+        $file_ext = isset($cache_config['file_ext']) ? $cache_config['file_ext'] : '.html';
+        $this->cache_file = $cache_config['cache_path'] . DS . $cache_config['key'] . $file_ext;
+        $this->expire_time = isset($cache_config ['expire_time']) ? $cache_config ['expire_time'] : 3600;
     }
 
     /**
@@ -31,8 +36,7 @@ class FileCache implements CacheInterface
      */
     function init()
     {
-        if (! file_exists($this->cache_file))
-        {
+        if (!file_exists($this->cache_file)) {
             Helper::mkfile($this->cache_file);
         }
     }
@@ -43,11 +47,10 @@ class FileCache implements CacheInterface
      * @param string $key
      * @return mixed
      */
-    function get( $key = '' )
+    function get($key = '')
     {
-        if(file_exists( $this->cache_file ))
-        {
-            return file_get_contents( $this->cache_file );
+        if (file_exists($this->cache_file)) {
+            return file_get_contents($this->cache_file);
         }
 
         return false;
@@ -60,19 +63,13 @@ class FileCache implements CacheInterface
      */
     function getExtime()
     {
-        if(! file_exists($this->cache_file))
-        {
+        if (!file_exists($this->cache_file)) {
             return false;
-        }
-
-        if( (time() - filemtime( $this->cache_file )) < $this->expire_time)
-        {
+        } elseif ((time() - filemtime($this->cache_file)) < $this->expire_time) {
             return true;
         }
-        else
-        {
-            return false;
-        }
+
+        return false;
     }
 
     /**
@@ -82,13 +79,11 @@ class FileCache implements CacheInterface
      * @param $value
      * @return mixed|void
      */
-    function set( $key, $value )
+    function set($key, $value)
     {
-        if (null == $key)
-        {
+        if (null == $key) {
             $key = $this->cache_file;
-            if (! file_exists($key))
-            {
+            if (!file_exists($key)) {
                 $this->init();
             }
         }
