@@ -3,12 +3,16 @@
  * @Auth: wonli <wonli@live.com>
  * Class RequestRedisCache
  */
+namespace cross\cache;
+
+use cross\i\CacheInterface;
+
 class RequestRedisCache extends RedisCache implements CacheInterface
 {
     function __construct($option)
     {
         parent::__construct($option);
-        $this->cache_key  = $option ['key'];
+        $this->cache_key = $option ['key'];
         $this->key_ttl = $option ['expire_time'];
     }
 
@@ -21,7 +25,7 @@ class RequestRedisCache extends RedisCache implements CacheInterface
      */
     function set($key, $value)
     {
-        $this->link->setex( $this->cache_key, $this->key_ttl, $value);
+        $this->link->setex($this->cache_key, $this->key_ttl, $value);
     }
 
     /**
@@ -31,7 +35,7 @@ class RequestRedisCache extends RedisCache implements CacheInterface
      */
     function getExtime()
     {
-        return $this->link->ttl( $this->cache_key ) > 0;
+        return $this->link->ttl($this->cache_key) > 0;
     }
 
     /**
@@ -40,12 +44,12 @@ class RequestRedisCache extends RedisCache implements CacheInterface
      * @param string $key
      * @return bool|mixed|string
      */
-    function get( $key = '' )
+    function get($key = '')
     {
-        if (! $key) {
+        if (!$key) {
             $key = $this->cache_key;
         }
 
-        return $this->link->get( $key );
+        return $this->link->get($key);
     }
 }
