@@ -3,25 +3,24 @@
  * @Auth: wonli <wonli@live.com>
  * Class FrameBase
  */
+namespace cross\core;
 
-class FrameBase
+use cross\exception\CoreException;
+use cross\lib\mcrypt\Mcrypt;
+use Exception;
+use ReflectionClass;
+
+class FrameBase extends Application
 {
     /**
-     * app配置
-     *
-     * @var array
-     */
-    protected $config;
-
-    /**
-     * 请求的参数列表
+     * 参数列表
      *
      * @var array
      */
     protected $params;
 
     /**
-     * action
+     * 方法名称
      *
      * @var string
      */
@@ -33,6 +32,20 @@ class FrameBase
      * @var string
      */
     protected $controller;
+
+    /**
+     * app配置
+     *
+     * @var Config
+     */
+    protected $config;
+
+    /**
+     * 缓存配置
+     *
+     * @var string
+     */
+    protected $cache_config;
 
     /**
      * url加密时用到的key
@@ -65,106 +78,21 @@ class FrameBase
      */
     function app_init()
     {
-        if(! $this->config) {
-            $this->setConfig( Dispatcher::getConfig() );
+        if (!$this->config) {
+            $this->config = parent::getConfig();
         }
 
-        if(! $this->controller) {
-            $this->setControllerName( Dispatcher::getController() );
+        if (!$this->controller) {
+            $this->controller = parent::getController();
         }
 
-        if(! $this->action) {
-            $this->setActionName( Dispatcher::getAction() );
+        if (!$this->action) {
+            $this->action = parent::getAction();
         }
 
-        if(! $this->params) {
-            $this->setParams( Dispatcher::getParams() );
+        if (!$this->params) {
+            $this->params = parent::getParams();
         }
-    }
-
-    /**
-     * 设置控制器名称
-     *
-     * @param $controller_name
-     */
-    public function setControllerName($controller_name)
-    {
-        $this->controller = $controller_name;
-    }
-
-    /**
-     * 取得控制器名称
-     *
-     * @return mixed
-     */
-    public function getControllerName() {
-        return $this->controller;
-    }
-
-    /**
-     * 设置Action
-     *
-     * @param $action_name
-     */
-    public function setActionName($action_name)
-    {
-        $this->action = $action_name;
-    }
-
-    /**
-     * 获取action
-     *
-     * @return string
-     */
-    public function getAction()
-    {
-        return $this->action;
-    }
-
-    /**
-     * 设置params
-     *
-     * @param $params
-     */
-    public function setParams($params)
-    {
-        $this->params = $params;
-    }
-
-    /**
-     * 取得url参数列表
-     *
-     * @param bool $strip 是否过滤html标签
-     * @return action|mixed
-     */
-    public function getParams( $strip = false )
-    {
-        if(true === $strip)
-        {
-            return Helper::strip_selected_tags($this->params);
-        }
-
-        return $this->params;
-    }
-
-    /**
-     * 设置全局配置
-     *
-     * @param $config
-     */
-    public function setConfig($config)
-    {
-         $this->config = $config;
-    }
-
-    /**
-     * 返回配置
-     *
-     * @return mixed
-     */
-    public function getConfig()
-    {
-        return $this->config;
     }
 
     /**
@@ -184,7 +112,7 @@ class FrameBase
      */
     public function getCacheConfig()
     {
-       return $this->cache_config;
+        return $this->cache_config;
     }
 
     /**
@@ -195,10 +123,10 @@ class FrameBase
      * @param int $exp 过期时间
      * @return bool
      */
-    protected function setAuth($key, $value, $exp=86400)
+    protected function setAuth($key, $value, $exp = 86400)
     {
         $auth_type = $this->config->get("sys", "auth");
-        return HttpAuth::factory( $auth_type )->set($key, $value, $exp);
+        return HttpAuth::factory($auth_type)->set($key, $value, $exp);
     }
 
     /**
@@ -211,7 +139,7 @@ class FrameBase
     protected function getAuth($key, $de = false)
     {
         $auth_type = $this->config->get("sys", "auth");
-        return HttpAuth::factory( $auth_type )->get($key, $de);
+        return HttpAuth::factory($auth_type)->get($key, $de);
     }
 
     /**
@@ -221,7 +149,7 @@ class FrameBase
      * @param string $type
      * @return bool|string
      */
-    protected function urlEncrypt($tex, $type="encode")
+    protected function urlEncrypt($tex, $type = "encode")
     {
         $key = $this->getUrlEncryptKey();
         return Helper::encode_params($tex, $key, $type);
@@ -242,10 +170,9 @@ class FrameBase
      */
     protected function getUrlEncryptKey()
     {
-        if (! $this->url_crypt_key)
-        {
+        if (!$this->url_crypt_key) {
             $url_crypt_key = $this->config->get('url', 'crypto_key');
-            if (! $url_crypt_key) {
+            if (!$url_crypt_key) {
                 $url_crypt_key = 'crossphp';
             }
 
@@ -261,17 +188,17 @@ class FrameBase
      * @param null $params
      * @return bool|string
      */
-    protected function sParams($params=null)
+    protected function sParams($params = null)
     {
         $url_type = $this->config->get('url', 'type');
 
         if (null === $params) {
             if ($url_type == 2) {
-                $params = current(array_keys( $this->params ));
+                $params = current(array_keys($this->params));
             } else {
                 $params = $this->params;
                 if (is_array($this->params)) {
-                    $params = current( array_values( $this->params ) );
+                    $params = current(array_values($this->params));
                 }
             }
         }
@@ -290,8 +217,8 @@ class FrameBase
             parse_str($decode_params_str, $result);
         } else {
             $result_array = explode($this->config->get('url', 'dot'), $decode_params_str);
-            $annotate = Dispatcher::getActionConfig();
-            $result = Dispatcher::combineParamsAnnotateConfig($result_array, $annotate['params']);
+            $annotate = parent::getActionConfig();
+            $result = parent::combineParamsAnnotateConfig($result_array, $annotate['params']);
         }
 
         return $result;
@@ -307,6 +234,7 @@ class FrameBase
     {
         $mcrypt = new Mcrypt;
         $_params = $mcrypt->enCode($params);
+
         return $_params[1];
     }
 
@@ -320,55 +248,8 @@ class FrameBase
     {
         $mcrypt = new Mcrypt;
         $_params = $mcrypt->deCode($params);
+
         return $_params;
-    }
-
-    /**
-     * 加载其他控制器
-     *
-     * @param $controller_name
-     * @param array $params
-     * @param bool $run_action
-     * <pre>
-     *  如果$run_action=true
-     *      $controller_name 支持 控制器:方法 的形式调用
-     * </pre>
-     * @return array|mixed|string
-     */
-    protected function loadController($controller_name, $params=array(), $run_action=false)
-    {
-        return Dispatcher::init( APP_NAME, $this->config )->run( $controller_name, $params, $run_action );
-    }
-
-    /**
-     * 加载指定Module
-     *
-     * @param $module_name
-     * @param string $params
-     * @return mixed
-     */
-    protected function loadModule( $module_name, $params = '' )
-    {
-        if (! isset(self::$moduleInstance[ $module_name ]))
-        {
-            $args = func_get_args();
-            if ($params != '' && count($args) > 2)
-            {
-                array_shift($args);
-                $params = $args;
-            }
-
-            if (false !== strpos( $module_name, "/"))
-            {
-                Loader::import("::modules/{$module_name}Module");
-                $module_name = end( explode("/", $module_name) );
-            }
-
-            $_module = ucfirst("{$module_name}Module");
-            self::$moduleInstance[ $module_name ] = new $_module( $params );
-        }
-
-        return self::$moduleInstance[ $module_name ];
     }
 
     /**
@@ -378,19 +259,17 @@ class FrameBase
      * @return mixed
      * @throws CoreException
      */
-    protected function loadObject( $objectInstance )
+    protected function loadObject($objectInstance)
     {
-        try
-        {
-            $obj = new ReflectionClass( $objectInstance );
-            if(! isset(self::$objectCache[ $obj->name ]))
-            {
-                self::$objectCache[ $obj->name ] = $objectInstance;
+        try {
+            $obj = new ReflectionClass($objectInstance);
+            if (!isset(self::$objectCache[$obj->name])) {
+                self::$objectCache[$obj->name] = $objectInstance;
             }
 
-            return self::$objectCache[ $obj->name ];
+            return self::$objectCache[$obj->name];
         } catch (Exception $e) {
-            throw new CoreException( 'cache module failed!' );
+            throw new CoreException('cache module failed!');
         }
     }
 
@@ -401,46 +280,20 @@ class FrameBase
      * @param null $controller_name
      * @return mixed
      */
-    protected function loadView($controller_name = null, $action = null)
+    protected function initView($controller_name = null, $action = null)
     {
-        if(null == $controller_name)
-        {
+        if (null == $controller_name) {
             $controller_name = $this->controller;
         }
 
-        $view_class_name = "{$controller_name}View";
+        $view_class_name = 'app\\' . APP_NAME . "\\views\\{$controller_name}View";
         $view = new $view_class_name;
 
-        if(null !== $action)
-        {
+        if (null !== $action) {
             return $view->$action();
         }
 
         return $view;
-    }
-
-    /**
-     * 运行app
-     *
-     * @param $act
-     * @return string
-     */
-    function run( $act )
-    {
-        $mr = new ReflectionClass($this);
-
-        $_before = "{$act}_before";
-        $_after = "{$act}_after";
-
-        ob_start();
-        if ( $mr->hasMethod($_before) ) {
-            echo $this->$_before();
-        }
-        echo $this->$act( );
-        if ( $mr->hasMethod($_after) ) {
-            echo $this->$_after();
-        }
-        return ob_get_clean();
     }
 
     /**
@@ -451,16 +304,17 @@ class FrameBase
      * @param string $type
      * @return array|string
      */
-    function result($ok = "1", $msg = "ok", $type="")
+    function result($ok = "1", $msg = "ok", $type = "")
     {
         $result = array();
 
         $result["status"] = $ok;
         $result["message"] = $msg;
 
-        if($type == "JSON") {
+        if ($type == "JSON") {
             $result = json_encode($result);
         }
+
         return $result;
     }
 
@@ -470,10 +324,9 @@ class FrameBase
      * @param $property
      * @return mixed|object|Response
      */
-    function __get( $property )
+    function __get($property)
     {
-        switch($property)
-        {
+        switch ($property) {
             case 'request' :
                 return $this->request = Request::getInstance();
 
@@ -481,7 +334,7 @@ class FrameBase
                 return $this->response = Response::getInstance();
 
             case 'view' :
-                return $this->view = $this->loadView();
+                return $this->view = $this->initView();
 
             default :
                 return null;
