@@ -3,6 +3,10 @@
  * @Auth: wonli <wonli@live.com>
  * Class Request
  */
+namespace cross\core;
+
+use cross\exception\FrontException;
+
 class Request
 {
     /**
@@ -42,7 +46,7 @@ class Request
      */
     public static function getInstance()
     {
-        if(! self::$instance ) {
+        if (!self::$instance) {
             self::$instance = new Request();
         }
 
@@ -61,28 +65,23 @@ class Request
             throw new FrontException('determine the entry script URL failed!!!');
         }
         $scriptName = basename($scriptName);
-        if (($_scriptName = $this->_SERVER('SCRIPT_NAME')) != null && basename($_scriptName) === $scriptName)
-        {
+        if (($_scriptName = $this->_SERVER('SCRIPT_NAME')) != null && basename($_scriptName) === $scriptName) {
             $this->_scriptUrl = $_scriptName;
-        }
-        elseif (($_scriptName = $this->_SERVER('PHP_SELF')) != null && basename($_scriptName) === $scriptName)
-        {
+        } elseif (($_scriptName = $this->_SERVER('PHP_SELF')) != null && basename($_scriptName) === $scriptName) {
             $this->_scriptUrl = $_scriptName;
-        }
-        elseif (($_scriptName = $this->_SERVER('ORIG_SCRIPT_NAME')) != null && basename($_scriptName) === $scriptName)
-        {
+        } elseif (($_scriptName = $this->_SERVER('ORIG_SCRIPT_NAME')) != null && basename(
+                                                                                     $_scriptName
+                                                                                 ) === $scriptName
+        ) {
             $this->_scriptUrl = $_scriptName;
-        }
-        elseif (($pos = strpos($this->_SERVER('PHP_SELF'), '/' . $scriptName)) !== false)
-        {
+        } elseif (($pos = strpos($this->_SERVER('PHP_SELF'), '/' . $scriptName)) !== false) {
             $this->_scriptUrl = substr($this->_SERVER('SCRIPT_NAME'), 0, $pos) . '/' . $scriptName;
         } elseif (($_documentRoot = $this->_SERVER('DOCUMENT_ROOT')) != null && ($_scriptName = $this->_SERVER(
-            'SCRIPT_FILENAME')) != null && strpos($_scriptName, $_documentRoot) === 0)
-        {
+                'SCRIPT_FILENAME'
+            )) != null && strpos($_scriptName, $_documentRoot) === 0
+        ) {
             $this->_scriptUrl = str_replace('\\', '/', str_replace($_documentRoot, '', $_scriptName));
-        }
-        else
-        {
+        } else {
             throw new FrontException('determine the entry script URL failed!!');
         }
     }
@@ -105,41 +104,44 @@ class Request
      */
     public function getScriptUrl()
     {
-        if (!$this->_scriptUrl) $this->_initScriptUrl();
+        if (!$this->_scriptUrl) {
+            $this->_initScriptUrl();
+        }
+
         return $this->_scriptUrl;
     }
 
-	/**
-	 * 设置基础路径
+    /**
+     * 设置基础路径
      *
-	 * @param  string $url 设置基础路径
-	 */
+     * @param  string $url 设置基础路径
+     */
     public function setBaseUrl($url)
     {
         $this->_baseUrl = $url;
     }
 
-	/**
-	 * 返回baseurl
+    /**
+     * 返回baseurl
      *
-	 * @param  boolean $absolute 是否返回带HOST的绝对路径
-	 * @return string 当前请求的url
-	 */
-	public function getBaseUrl($absolute = false)
+     * @param  boolean $absolute 是否返回带HOST的绝对路径
+     * @return string 当前请求的url
+     */
+    public function getBaseUrl($absolute = false)
     {
-		if ($this->_baseUrl === null) $this->_baseUrl = rtrim(dirname($this->getScriptUrl()), '\\/.');
-		return $absolute ? $this->getHostInfo() . $this->_baseUrl : $this->_baseUrl;
-	}
+        if ($this->_baseUrl === null) $this->_baseUrl = rtrim(dirname($this->getScriptUrl()), '\\/.');
+        return $absolute ? $this->getHostInfo() . $this->_baseUrl : $this->_baseUrl;
+    }
 
-	/**
-	 * 返回当前脚本文件名
+    /**
+     * 返回当前脚本文件名
      *
-	 * @return string 当前请求的脚本名
-	 */
+     * @return string 当前请求的脚本名
+     */
     public function getScriptName()
     {
         $s = explode("/", $this->_SERVER('SCRIPT_NAME'));
-        return end( $s );
+        return end($s);
     }
 
     /**
@@ -149,7 +151,7 @@ class Request
      */
     public function getIndexName()
     {
-        if($this->_indexName === null) return $this->getScriptName();
+        if ($this->_indexName === null) return $this->getScriptName();
         return $this->_indexName;
     }
 
@@ -161,16 +163,16 @@ class Request
         $this->_indexName = $index_name;
     }
 
-	/**
-	 * 取得Host信息
+    /**
+     * 取得Host信息
      *
-	 * @return null
-	 */
-	public function getHostInfo()
+     * @return null
+     */
+    public function getHostInfo()
     {
-        if(!$this->_hostInfo) $this->_initHostInfo();
-		return $this->_hostInfo;
-	}
+        if (!$this->_hostInfo) $this->_initHostInfo();
+        return $this->_hostInfo;
+    }
 
     /**
      * 设置Host信息
@@ -178,39 +180,34 @@ class Request
      * @throws FrontException
      * @return null
      */
-	private function _initHostInfo()
+    private function _initHostInfo()
     {
-		if(PHP_SAPI === 'cli')
-        {
-            return "cli";
+        if (PHP_SAPI === 'cli') {
+            return 'cli';
         }
 
-		$http = $this->_SERVER("HTTPS") == 'on'?'https':'http';
-		if (($httpHost = $this->_SERVER('HTTP_HOST')) != null)
-        {
+        $http = $this->_SERVER("HTTPS") == 'on' ? 'https' : 'http';
+        if (($httpHost = $this->_SERVER('HTTP_HOST')) != null) {
             $this->_hostInfo = $http . '://' . $httpHost;
-        }
-		elseif (($httpHost = $this->_SERVER('SERVER_NAME')) != null)
-        {
-			$this->_hostInfo = $http . '://' . $httpHost;
-			if (($port = $this->getServerPort()) != null) $this->_hostInfo .= ':' . $port;
-		}
-        else
-        {
+        } elseif (($httpHost = $this->_SERVER('SERVER_NAME')) != null) {
+            $this->_hostInfo = $http . '://' . $httpHost;
+            if (($port = $this->getServerPort()) != null) $this->_hostInfo .= ':' . $port;
+        } else {
             throw new FrontException('determine the entry script URL failed!!');
         }
-	}
+    }
 
     /**
      * 取得服务器端口
      *
      * @return int 当前服务器端口号
      */
-	public function getServerPort()
+    public function getServerPort()
     {
-        $_default = $this->_SERVER("HTTPS") == 'on'?443:80;
-		return $_default;
-	}
+        $_default = $this->_SERVER("HTTPS") == 'on' ? 443 : 80;
+
+        return $_default;
+    }
 
     /**
      * 当前scriptFile的路径
@@ -223,6 +220,7 @@ class Request
         if (($scriptName = $this->_SERVER('SCRIPT_FILENAME')) == null) {
             throw new FrontException('determine the entry script URL failed!!!');
         }
+
         return realpath(dirname($scriptName));
     }
 
@@ -230,9 +228,9 @@ class Request
      * @return string
      */
     public function getUserHost()
-	{
-		return $this->_SERVER('REMOTE_HOST');
-	}
+    {
+        return $this->_SERVER('REMOTE_HOST');
+    }
 
     /**
      * 按type返回url请求
@@ -253,9 +251,9 @@ class Request
      * @return string
      */
     public function getQueryString()
-	{
-		return $this->_SERVER('QUERY_STRING');
-	}
+    {
+        return $this->_SERVER('QUERY_STRING');
+    }
 
     /**
      * 是否是PUT请求
@@ -263,19 +261,22 @@ class Request
      * @return bool
      */
     public function isPutRequest()
-	{
-		return ( $this->_SERVER('REQUEST_METHOD') && !strcasecmp($this->_SERVER('REQUEST_METHOD'), 'PUT') ) || $this->isPutViaPostRequest();
-	}
+    {
+        return ($this->_SERVER('REQUEST_METHOD') && !strcasecmp(
+                $this->_SERVER('REQUEST_METHOD'),
+                'PUT'
+            )) || $this->isPutViaPostRequest();
+    }
 
     /**
-   	 * 判断一个链接是否为post请求
+     * 判断一个链接是否为post请求
      *
-   	 * @return boolean
-   	 */
-   	public function isPostRequest()
-   	{
+     * @return boolean
+     */
+    public function isPostRequest()
+    {
         return $this->_SERVER('REQUEST_METHOD') && !strcasecmp($this->_SERVER('REQUEST_METHOD'), 'POST');
-   	}
+    }
 
     /**
      * 判断请求类型是否为get
@@ -303,9 +304,9 @@ class Request
      * @return bool
      */
     protected function isPutViaPostRequest()
-	{
-		return isset($_POST['_method']) && !strcasecmp($_POST['_method'], 'PUT');
-	}
+    {
+        return isset($_POST['_method']) && !strcasecmp($_POST['_method'], 'PUT');
+    }
 
     /**
      * 是否是ajax请求
@@ -313,9 +314,9 @@ class Request
      * @return bool
      */
     public function isAjaxRequest()
-	{
-		return $this->_SERVER('HTTP_X_REQUESTED_WITH')==='XMLHttpRequest';
-	}
+    {
+        return $this->_SERVER('HTTP_X_REQUESTED_WITH') === 'XMLHttpRequest';
+    }
 
     /**
      * 是否是flash请求
@@ -323,9 +324,14 @@ class Request
      * @return bool
      */
     public function isFlashRequest()
-	{
-		return  stripos($this->_SERVER('HTTP_USER_AGENT'),'Shockwave')!==false || stripos($this->_SERVER('HTTP_USER_AGENT'),'Flash')!==false;
-	}
+    {
+        return stripos($this->_SERVER('HTTP_USER_AGENT'), 'Shockwave') !== false || stripos(
+                                                                                        $this->_SERVER(
+                                                                                            'HTTP_USER_AGENT'
+                                                                                        ),
+                                                                                        'Flash'
+                                                                                    ) !== false;
+    }
 
     /**
      * HTTP_REFERER;
@@ -333,32 +339,32 @@ class Request
      * @return string
      */
     public function getUrlReferrer()
-	{
-		return $this->_SERVER('HTTP_REFERER');
-	}
+    {
+        return $this->_SERVER('HTTP_REFERER');
+    }
 
     /**
      * @return string userAgent
      */
     public function getUserAgent()
-	{
-		return $this->_SERVER('HTTP_USER_AGENT');
-	}
+    {
+        return $this->_SERVER('HTTP_USER_AGENT');
+    }
 
     /**
      * @return string userIP
      */
     public function getUserHostAddress()
-	{
-		return $this->_SERVER('REMOTE_ADDR') !== ''?$this->_SERVER('REMOTE_ADDR'):'127.0.0.1';
-	}
+    {
+        return $this->_SERVER('REMOTE_ADDR') !== '' ? $this->_SERVER('REMOTE_ADDR') : '127.0.0.1';
+    }
 
     /**
      * @return string ACCEPT TYPE
      */
     public function getAcceptTypes()
-	{
-		return $this->_SERVER('HTTP_ACCEPT');
-	}
+    {
+        return $this->_SERVER('HTTP_ACCEPT');
+    }
 }
 
