@@ -3,6 +3,9 @@
  * @Auth: wonli <wonli@live.com>
  * Rest.php
  */
+namespace cross\core;
+
+use Closure;
 
 class Rest
 {
@@ -29,13 +32,13 @@ class Rest
     /**
      * 初始化request
      */
-    private function __construct( $config )
+    private function __construct($config)
     {
         $this->request = Request::getInstance();
         $this->config = $config;
 
         $url_type = $this->config->get('url', 'type');
-        $this->request_string = $this->request->getUrlRequest( $url_type );
+        $this->request_string = $this->request->getUrlRequest($url_type);
     }
 
     /**
@@ -44,10 +47,10 @@ class Rest
      * @param $config
      * @return Rest
      */
-    static function getInstance( $config )
+    static function getInstance($config)
     {
-        if(! self::$instance) {
-            self::$instance = new Rest( $config );
+        if (!self::$instance) {
+            self::$instance = new Rest($config);
         }
 
         return self::$instance;
@@ -59,18 +62,17 @@ class Rest
      * @param $request_url
      * @param callable $process_func
      */
-    function get( $request_url, Closure $process_func )
+    function get($request_url, Closure $process_func)
     {
-        if( true !== $this->request->isGetRequest() ) {
+        if (true !== $this->request->isGetRequest()) {
             return;
         }
 
         $params = array();
-        if ( true === $this->checkRequest( $request_url, $params ))
-        {
+        if (true === $this->checkRequest($request_url, $params)) {
             $rep = call_user_func_array($process_func, $params);
             if (null != $rep) {
-                $this->response( $rep );
+                $this->response($rep);
             }
         }
     }
@@ -81,9 +83,9 @@ class Rest
      * @param $request_url
      * @param callable $process_func
      */
-    function post( $request_url, Closure $process_func )
+    function post($request_url, Closure $process_func)
     {
-        if ( true !== $this->request->isPostRequest() ) {
+        if (true !== $this->request->isPostRequest()) {
             return;
         }
 
@@ -94,7 +96,7 @@ class Rest
         $data = file_get_contents("php://input");
         $rep = call_user_func($process_func, $data);
         if (null != $rep) {
-            $this->response( $rep );
+            $this->response($rep);
         }
     }
 
@@ -104,7 +106,7 @@ class Rest
      * @param $request_url
      * @param callable $process_func
      */
-    function put( $request_url, Closure $process_func )
+    function put($request_url, Closure $process_func)
     {
         if (true !== $this->request->isPutRequest()) {
             return;
@@ -117,7 +119,7 @@ class Rest
         $data = file_get_contents("php://input");
         $rep = call_user_func($process_func, $data);
         if (null != $rep) {
-            $this->response( $rep );
+            $this->response($rep);
         }
     }
 
@@ -127,18 +129,17 @@ class Rest
      * @param $request_url
      * @param callable $process_func
      */
-    function delete( $request_url, Closure $process_func )
+    function delete($request_url, Closure $process_func)
     {
         if (true !== $this->request->isDeleteRequest()) {
             return;
         }
 
         $params = array();
-        if ( true === $this->checkRequest( $request_url, $params ))
-        {
+        if (true === $this->checkRequest($request_url, $params)) {
             $rep = call_user_func_array($process_func, $params);
             if (null != $rep) {
-                $this->response( $rep );
+                $this->response($rep);
             }
         }
     }
@@ -150,38 +151,33 @@ class Rest
      * @param $params
      * @return bool
      */
-    function checkRequest( $request_url, & $params )
+    function checkRequest($request_url, & $params)
     {
         $url_dot = $this->config->get("url", "dot");
         $params_key = array();
         $params_value = array();
 
-        if (($request_url == '' && $this->request_string != '') || ($request_url == '/' && $this->request_string != '/'))
-        {
+        if (($request_url == '' && $this->request_string != '') || ($request_url == '/' && $this->request_string != '/')) {
             return false;
         }
 
-        if (false !== strpos($request_url, "{$url_dot}:"))
-        {
-            $params_start = strpos( $request_url, "{$url_dot}:" );
-            $params_key = array_filter( explode("{$url_dot}:", substr( $request_url, $params_start )));
-            $request_url = substr( $request_url, 0, $params_start );
+        if (false !== strpos($request_url, "{$url_dot}:")) {
+            $params_start = strpos($request_url, "{$url_dot}:");
+            $params_key = array_filter(explode("{$url_dot}:", substr($request_url, $params_start)));
+            $request_url = substr($request_url, 0, $params_start);
         }
 
-        if (0 === strncasecmp($this->request_string, $request_url, strlen($request_url)))
-        {
-            if(! empty($params_key))
-            {
-                $params_value = array_filter( explode("{$url_dot}", substr($this->request_string, strlen($request_url))));
+        if (0 === strncasecmp($this->request_string, $request_url, strlen($request_url))) {
+            if (!empty($params_key)) {
+                $params_value = array_filter(
+                    explode("{$url_dot}", substr($this->request_string, strlen($request_url)))
+                );
             }
-        }
-        else
-        {
+        } else {
             return false;
         }
 
-        if (! empty($params_value) && count($params_key) == count($params_value))
-        {
+        if (!empty($params_value) && count($params_key) == count($params_value)) {
             $params = array_combine($params_key, $params_value);
         }
 
@@ -193,7 +189,7 @@ class Rest
      *
      * @param $rep
      */
-    function response( $rep )
+    function response($rep)
     {
         Response::getInstance()->output($rep);
     }
