@@ -4,13 +4,19 @@
  * @Auth: wonli <wonli@live.com>
  * Class FrontException
  */
+namespace cross\exception;
+
+use cross\core\Response;
+use exception;
+
 class FrontException extends CrossException
 {
-    function error_handler (exception $e)
+    function error_handler(exception $e)
     {
         $cp_error = $this->cp_exception_source($e);
         $code = $e->getCode() ? $e->getCode() : 200;
-        return Response::getInstance()->set_response_status($code)->display( $cp_error, CP_PATH.'exception/_tpl/fronterror.php' );
+
+        return Response::getInstance()->set_response_status($code)->display($cp_error, CP_PATH . 'exception/_tpl/fronterror.php');
     }
 }
 
