@@ -3,6 +3,9 @@
  * @Auth: wonli <wonli@live.com>
  * UploadImages.php
  */
+namespace cross\lib\images;
+
+use Exception;
 
 class UploadImages
 {
@@ -63,16 +66,16 @@ class UploadImages
      * @var array
      */
     protected $status_code = array(
-        -8  =>  '文件上传失败',
-        -4  =>  '大小超出限制',
-        -3  =>  '不允许上传的类型',
-         0  =>  '上传成功',
-         1  =>  '上传的文件超过了 php.ini 中 upload_max_filesize 选项限制的值',
-         2  =>  '上传文件的大小超过了 HTML 表单中 MAX_FILE_SIZE 选项指定的值',
-         3  =>  '文件只有部分被上传',
-         4  =>  '没有文件被上传',
-         6  =>  '找不到临时文件夹',
-         7  =>  '文件写入失败',
+        -8 => '文件上传失败',
+        -4 => '大小超出限制',
+        -3 => '不允许上传的类型',
+        0 => '上传成功',
+        1 => '上传的文件超过了 php.ini 中 upload_max_filesize 选项限制的值',
+        2 => '上传文件的大小超过了 HTML 表单中 MAX_FILE_SIZE 选项指定的值',
+        3 => '文件只有部分被上传',
+        4 => '没有文件被上传',
+        6 => '找不到临时文件夹',
+        7 => '文件写入失败',
     );
 
     function __construct($upload_file_name, $save_name = '')
@@ -95,9 +98,10 @@ class UploadImages
      * @param $size
      * @return $this
      */
-    protected function set_allow_size( $size )
+    protected function set_allow_size($size)
     {
         $this->allow_size = $size;
+
         return $this;
     }
 
@@ -106,7 +110,7 @@ class UploadImages
      *
      * @return int
      */
-    protected function get_allow_size( )
+    protected function get_allow_size()
     {
         return $this->allow_size;
     }
@@ -117,9 +121,10 @@ class UploadImages
      * @param string $path
      * @return $this
      */
-    public function set_save_path( $path = '' )
+    public function set_save_path($path = '')
     {
         $this->save_path = $path;
+
         return $this;
     }
 
@@ -130,11 +135,11 @@ class UploadImages
      */
     protected function get_save_path()
     {
-        if (! empty($this->save_path) && ! is_dir($this->save_path)) {
+        if (!empty($this->save_path) && !is_dir($this->save_path)) {
             mkdir($this->save_path, 755, true);
         }
 
-        return rtrim($this->save_path, DIRECTORY_SEPARATOR).DIRECTORY_SEPARATOR;
+        return rtrim($this->save_path, DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR;
     }
 
     /**
@@ -156,6 +161,7 @@ class UploadImages
     public function set_allow_type(array $type)
     {
         $this->allow_type = $type;
+
         return $this;
     }
 
@@ -167,20 +173,19 @@ class UploadImages
      */
     protected function get_base_info()
     {
-        if (! $this->upload_file_base_info)
-        {
-            if (isset($_FILES[$this->upload_file_name]))
-            {
+        if (!$this->upload_file_base_info) {
+            if (isset($_FILES[$this->upload_file_name])) {
                 $base_info = $_FILES[$this->upload_file_name];
                 if ($base_info['error'] !== 0) {
-                    throw new Exception( $base_info['error'] );
-                } else {
+                    throw new Exception($base_info['error']);
+                }
+                else {
                     $extend_info = $this->get_image_info($base_info ['tmp_name']);
-                    if (! empty($extend_info)) {
+                    if (!empty($extend_info)) {
                         $base_info = array_merge($base_info, $extend_info);
                     }
                 }
-                $this->set_base_info( $base_info );
+                $this->set_base_info($base_info);
             }
         }
 
@@ -192,7 +197,7 @@ class UploadImages
      *
      * @param $upload_file_base_info
      */
-    protected function set_base_info( $upload_file_base_info )
+    protected function set_base_info($upload_file_base_info)
     {
         $this->upload_file_base_info = $upload_file_base_info;
     }
@@ -205,7 +210,8 @@ class UploadImages
     protected function get_save_file_name()
     {
         $base_file = $this->get_base_info();
-        return $this->save_name.$base_file['ext'];
+
+        return $this->save_name . $base_file['ext'];
     }
 
     /**
@@ -233,7 +239,7 @@ class UploadImages
      *
      * @return array
      */
-    protected function move_upload_file( )
+    protected function move_upload_file()
     {
         try {
             $base_info = $this->get_base_info();
@@ -244,9 +250,10 @@ class UploadImages
         $check_ret = $this->check();
         if (true !== $check_ret) {
             return $this->result($check_ret);
-        } else {
+        }
+        else {
             $saved_file_path = $this->get_save_file_full_path();
-            if ( move_uploaded_file($base_info['tmp_name'], $saved_file_path) ) {
+            if (move_uploaded_file($base_info['tmp_name'], $saved_file_path)) {
                 return $this->result(0, $saved_file_path);
             }
         }
@@ -259,12 +266,11 @@ class UploadImages
      */
     public function save()
     {
-        $ret = $this->move_upload_file( );
-        if ($ret['status'] === 0)
-        {
+        $ret = $this->move_upload_file();
+        if ($ret['status'] === 0) {
             return $this->result('ok', array(
-                'url'   =>  $this->get_save_file_name(),
-                'path'  =>  $ret['message'],
+                'url' => $this->get_save_file_name(),
+                'path' => $ret['message'],
             ));
         }
 
@@ -278,23 +284,20 @@ class UploadImages
      * @param bool $save_ori_images
      * @return array
      */
-    public function thumb( $thumb_config = array(), $save_ori_images = true )
+    public function thumb($thumb_config = array(), $save_ori_images = true)
     {
-        $ret = $this->move_upload_file( );
-        if ($ret['status'] === 0)
-        {
+        $ret = $this->move_upload_file();
+        if ($ret['status'] === 0) {
             $result_file_url['ori'] = $this->get_save_file_name();
             //生成缩略图
-            if (! empty($thumb_config))
-            {
+            if (!empty($thumb_config)) {
                 $ori_images_path = $ret['message'];
-                if (! is_array($thumb_config))
-                {
+                if (!is_array($thumb_config)) {
                     $thumb_config = array($thumb_config);
                 }
 
-                $result_file_url['thumb'] = $this->make_thumb( $ori_images_path, $thumb_config );
-                if (! $save_ori_images) {
+                $result_file_url['thumb'] = $this->make_thumb($ori_images_path, $thumb_config);
+                if (!$save_ori_images) {
                     unlink($ori_images_path);
                     unset($result_file_url['ori']);
                 }
@@ -316,20 +319,17 @@ class UploadImages
      */
     public function cut($thumb_config, $coordinate = array(), $save_ori_images = true)
     {
-        $ret = $this->move_upload_file( );
-        if ($ret['status'] === 0)
-        {
+        $ret = $this->move_upload_file();
+        if ($ret['status'] === 0) {
             $result_file_url['ori'] = $this->get_save_file_name();
-            if (! empty($thumb_config))
-            {
+            if (!empty($thumb_config)) {
                 $ori_images_path = $ret['message'];
-                if (! is_array($thumb_config))
-                {
+                if (!is_array($thumb_config)) {
                     $thumb_config = array($thumb_config);
                 }
 
-                $result_file_url['thumb'] = $this->make_cut_thumb( $ori_images_path, $thumb_config, $coordinate );
-                if (! $save_ori_images) {
+                $result_file_url['thumb'] = $this->make_cut_thumb($ori_images_path, $thumb_config, $coordinate);
+                if (!$save_ori_images) {
                     unlink($ori_images_path);
                     unset($result_file_url['ori']);
                 }
@@ -352,16 +352,16 @@ class UploadImages
     protected function make_cut_thumb($images_path, $thumb_size_config, $coordinate)
     {
         $result = array();
-        if (! $images_path || ! $thumb_size_config) {
+        if (!$images_path || !$thumb_size_config) {
             return $result;
         }
 
         $cut = new ImageCut($images_path);
-        foreach($thumb_size_config as $val)
-        {
+        foreach ($thumb_size_config as $val) {
             if (false !== strpos($val, 'x')) {
                 list($width, $height) = explode('x', $val);
-            } else {
+            }
+            else {
                 $width = $height = $val;
             }
 
@@ -383,16 +383,16 @@ class UploadImages
     protected function make_thumb($images_path, $thumb_size_config)
     {
         $result = array();
-        if (! $images_path || ! $thumb_size_config) {
+        if (!$images_path || !$thumb_size_config) {
             return $result;
         }
 
         $Thumb = new ImageThumb($images_path);
-        foreach($thumb_size_config as $val)
-        {
+        foreach ($thumb_size_config as $val) {
             if (false !== strpos($val, 'x')) {
                 list($width, $height) = explode('x', $val);
-            } else {
+            }
+            else {
                 $width = $height = $val;
             }
 
@@ -410,7 +410,7 @@ class UploadImages
     {
         $base_info = $this->get_base_info();
         $allow_type = $this->get_allow_type();
-        if (! in_array($base_info['file_type'], $allow_type)) {
+        if (!in_array($base_info['file_type'], $allow_type)) {
             return -3;
         }
 
@@ -444,7 +444,7 @@ class UploadImages
         $save_path = $this->get_save_path();
         $save_name = $this->get_save_file_name();
 
-        return $save_path.$save_name;
+        return $save_path . $save_name;
     }
 
     /**
@@ -456,22 +456,23 @@ class UploadImages
     protected function get_image_info($images)
     {
         $image_info = getimagesize($images);
-        if (false !== $image_info)
-        {
-            $image_ext  = strtolower(image_type_to_extension($image_info[2]));
+        if (false !== $image_info) {
+            $image_ext = strtolower(image_type_to_extension($image_info[2]));
             $image_type = substr($image_ext, 1);
             $image_size = filesize($images);
 
             $info = array(
-                'width'		=>  $image_info[0],
-                'height'	=>  $image_info[1],
-                'ext'       =>  $image_ext,
-                'file_type' =>  $image_type,
-                'size'		=>  $image_size,
-                'mime'      =>  $image_info['mime'],
+                'width' => $image_info[0],
+                'height' => $image_info[1],
+                'ext' => $image_ext,
+                'file_type' => $image_type,
+                'size' => $image_size,
+                'mime' => $image_info['mime'],
             );
+
             return $info;
-        }else {
+        }
+        else {
             return false;
         }
     }
@@ -483,12 +484,11 @@ class UploadImages
      * @param string $message
      * @return array
      */
-    protected function result($code, $message='')
+    protected function result($code, $message = '')
     {
-        $result = array('status'=>0, 'message'=>'ok');
+        $result = array('status' => 0, 'message' => 'ok');
         $result['status'] = $code;
-        if ('' === $message)
-        {
+        if ('' === $message) {
             $message = $this->status_code[$code];
         }
         $result['message'] = $message;
