@@ -2,37 +2,39 @@
 /**
  * @Author: wonli <wonli@live.com>
  */
+namespace cross\model;
+
+use cross\cache\RedisCache;
+use cross\exception\CoreException;
+
 class CoreModel
 {
     static function factory($link_type, $link_params)
     {
-        switch( strtolower($link_type) )
-        {
+        switch (strtolower($link_type)) {
             case 'mysql' :
 
-                $port = isset($link_params['port'])?$link_params['port']:3306;
-                $char_set = isset($link_params['charset'])?$link_params['charset']:'utf8';
-                $options = isset($link_params['options'])?$link_params['options']:array();
+                $port = isset($link_params['port']) ? $link_params['port'] : 3306;
+                $char_set = isset($link_params['charset']) ? $link_params['charset'] : 'utf8';
+                $options = isset($link_params['options']) ? $link_params['options'] : array();
 
-                if (strcasecmp(PHP_OS, 'linux') == 0 && !empty($link_params['unix_socket']))
-                {
+                if (strcasecmp(PHP_OS, 'linux') == 0 && !empty($link_params['unix_socket'])) {
                     $dsn = "mysql:dbname={$link_params['name']};unix_socket={$link_params['unix_socket']}";
                 }
-                else
-                {
+                else {
                     $dsn = "mysql:host={$link_params['host']};dbname={$link_params['name']};port={$port};charset={$char_set}";
                 }
 
                 return MysqlModel::getInstance($dsn, $link_params["user"], $link_params["pass"], $options);
 
             case 'mongo':
-                return new MongoModel( $link_params );
+                return new MongoModel($link_params);
 
             case 'redis':
-                return new RedisCache( $link_params );
+                return new RedisCache($link_params);
 
             case 'couch':
-                return new CouchModel( $link_params );
+                return new CouchModel($link_params);
 
             default:
                 throw new CoreException("不支持的数据库扩展!");
