@@ -3,11 +3,18 @@
  * @Auth wonli <wonli@live.com>
  * Class MysqlModel
  */
+namespace cross\model;
 
-class MysqlModel implements SqlInterface
+use cross\exception\CoreException;
+use cross\exception\FrontException;
+use Exception;
+use PDO;
+use PDOException;
+
+class MysqlModel
 {
     /**
-     * @var $stmt
+     * @var PDO::PDOStatement
      */
     public $stmt;
 
@@ -15,6 +22,11 @@ class MysqlModel implements SqlInterface
      * @var PDO
      */
     public $pdo;
+
+    /**
+     * @var string
+     */
+    protected $sql;
 
     /**
      * 数据库连接实例
@@ -47,13 +59,13 @@ class MysqlModel implements SqlInterface
     /**
      * 创建数据库连接
      *
-     * @param $dsn string dsn
-     * @param $user string 数据库用户名
-     * @param $password string 数据库密码
+     * @param string $dsn dsn
+     * @param string $user 数据库用户名
+     * @param string $password 数据库密码
      * @param array $options
      * @throws CoreException
      */
-    private function __construct( $dsn, $user, $password, $options = array() )
+    private function __construct($dsn, $user, $password, $options = array())
     {
         try {
 
@@ -62,25 +74,23 @@ class MysqlModel implements SqlInterface
             }
 
             $this->pdo = new PDO($dsn, $user, $password, $options);
-        } catch(Exception $e) {
-            throw new CoreException($e->getMessage().' line:'.$e->getLine().' '.$e->getFile());
+        } catch (Exception $e) {
+            throw new CoreException($e->getMessage() . ' line:' . $e->getLine() . ' ' . $e->getFile());
         }
     }
 
     /**
      * @see MysqlModel::__construct
-     *
-     * @param $dsn
-     * @param $user
-     * @param $password
+     * @param string $dsn
+     * @param string $user
+     * @param string $password
      * @return mixed
      */
-    static function getInstance( $dsn, $user, $password )
+    static function getInstance($dsn, $user, $password)
     {
         //同时建立多个数据库连接
-        $_instance =  md5($dsn.$user.$password);
-        if(! isset( self::$instance[ $_instance ]) )
-        {
+        $_instance = md5($dsn . $user . $password);
+        if (!isset(self::$instance[$_instance])) {
             self::$instance [$_instance] = new self($dsn, $user, $password);
         }
 
@@ -89,10 +99,9 @@ class MysqlModel implements SqlInterface
 
     /**
      * @see MysqlModel::prepare_getone
-     *
-     * @param  $table string
-     * @param  $fields string
-     * @param  $where array
+     * @param string $table
+     * @param string $fields
+     * @param string|array $where
      * @return mixed
      */
     function get($table, $fields, $where)
@@ -102,24 +111,22 @@ class MysqlModel implements SqlInterface
 
     /**
      * @see MysqlModel::prepare_get_all
-     *
-     * @param $table
-     * @param $fields
-     * @param null $where
+     * @param string $table
+     * @param string $fields
+     * @param string|array $where
      * @param int $order
      * @param int $group_by
      * @return array
      */
-    function getAll($table, $fields, $where=null, $order=1, $group_by = 1)
+    function getAll($table, $fields, $where = null, $order = 1, $group_by = 1)
     {
         return $this->prepare_get_all($table, $fields, $where, $order, $group_by);
     }
 
     /**
      * @see MysqlModel::prepare_insert
-     *
-     * @param $table
-     * @param $data
+     * @param string $table
+     * @param string $data
      * @param bool $multi
      * @param array $insert_data
      * @return array|bool|mixed
@@ -131,27 +138,25 @@ class MysqlModel implements SqlInterface
 
     /**
      * @see MysqlModel::prepare_find
-     *
      * @param $table
      * @param $fields
      * @param $where
      * @param int $order
      * @param array $page
-     * @param string $group_by
+     * @param int|string $group_by
      * @return array|mixed
      */
-    function find($table, $fields, $where, $order = 1, & $page = array('p'=>1, 'limit'=>50), $group_by = 1)
+    function find($table, $fields, $where, $order = 1, & $page = array('p' => 1, 'limit' => 50), $group_by = 1)
     {
         return $this->prepare_find($table, $fields, $where, $order, $page, $group_by);
     }
 
     /**
      * @see MysqlModel::prepare_update
-     *
-     * @param $table
-     * @param $data
-     * @param $where
-     * @return $this|array|mixed|string
+     * @param string $table
+     * @param string $data
+     * @param string|array $where
+     * @return $this|array|string
      */
     function update($table, $data, $where)
     {
@@ -160,13 +165,12 @@ class MysqlModel implements SqlInterface
 
     /**
      * @see MysqlModel::prepare_delete
-     *
-     * @param $table
-     * @param $where
+     * @param string $table
+     * @param string $where
      * @param bool $multi
      * @return bool|mixed
      */
-    function del($table, $where, $multi=false)
+    function del($table, $where, $multi = false)
     {
         return $this->prepare_delete($table, $where, $multi);
     }
@@ -174,92 +178,90 @@ class MysqlModel implements SqlInterface
     /**
      * 执行一条SQL 语句 并返回结果
      *
-     * @param $sql
+     * @param string $sql
      * @param int $model
      * @throws CoreException
      * @return mixed
      */
     function fetchOne($sql, $model = PDO::FETCH_ASSOC)
     {
-        try
-        {
-            $data = $this->pdo->query( $sql );
-            return $data->fetch( $model );
-        } catch( Exception $e ) {
-            throw new CoreException( $e->getMessage() );
+        try {
+            $data = $this->pdo->query($sql);
+
+            return $data->fetch($model);
+        } catch (Exception $e) {
+            throw new CoreException($e->getMessage());
         }
     }
 
     /**
      * 执行sql 并返回所有结果
      *
-     * @param $sql
+     * @param string $sql
      * @param int $model
      * @throws CoreException
      * @return array
      */
     function fetchAll($sql, $model = PDO::FETCH_ASSOC)
     {
-        try
-        {
-            $data = $this->pdo->query( $sql );
-            return $data->fetchAll( $model );
-        } catch ( Exception $e ) {
-            throw new CoreException( $e->getMessage() );
+        try {
+            $data = $this->pdo->query($sql);
+            return $data->fetchAll($model);
+        } catch (Exception $e) {
+            throw new CoreException($e->getMessage());
         }
     }
 
     /**
      * 执行sql 用于无返回值的操作
      *
-     * @param $sql
+     * @param string $sql
      * @return int
      * @throws CoreException
      */
     function execute($sql)
     {
-        try
-        {
+        try {
             return $this->pdo->exec($sql);
-        } catch ( Exception $e ) {
-            throw new CoreException( $e->getMessage() );
+        } catch (Exception $e) {
+            throw new CoreException($e->getMessage());
         }
     }
 
     /**
      * 参数绑定
      *
-     * @param $statement
+     * @param string $statement
      * @param array $params
      * @return $this
      * @throws CoreException
      */
     public function prepare($statement, $params = array(PDO::ATTR_CURSOR => PDO::CURSOR_FWDONLY))
     {
-        try
-        {
+        try {
             $this->stmt = $this->pdo->prepare($statement, $params);
+
             return $this;
         } catch (PDOException $e) {
-            throw new CoreException( $e->getMessage() );
+            throw new CoreException($e->getMessage());
         }
     }
 
     /**
      * 执行参数绑定
      *
-     * @param null $args
+     * @param array $args
      * @return $this
      * @throws CoreException
      */
-    public function exec($args=null)
+    public function exec($args = array())
     {
-        try
-        {
+        try {
             $this->stmt->execute($args);
+
             return $this;
         } catch (PDOException $e) {
-            throw new CoreException( $e->getMessage() );
+            throw new CoreException($e->getMessage());
         }
     }
 
@@ -267,26 +269,26 @@ class MysqlModel implements SqlInterface
      * 返回参数绑定结果
      *
      * @param bool $_fetchAll
-     * @param $result_type
+     * @param int $result_type
      * @return mixed
      * @throws CoreException
      */
-    public function stmt_fetch($_fetchAll=false, $result_type = PDO::FETCH_ASSOC)
+    public function stmt_fetch($_fetchAll = false, $result_type = PDO::FETCH_ASSOC)
     {
-        if(! $this->stmt) throw new CoreException("stmt init failed!");
-        if(true === $_fetchAll)
-        {
+        if (!$this->stmt) throw new CoreException("stmt init failed!");
+        if (true === $_fetchAll) {
             return $this->stmt->fetchAll($result_type);
         }
+
         return $this->stmt->fetch($result_type);
     }
 
     /**
      * prepare方式单条查询
      *
-     * @param $table
-     * @param $fields
-     * @param $where
+     * @param string $table
+     * @param string $fields
+     * @param string|array $where
      * @return mixed
      */
     function prepare_getone($table, $fields, $where)
@@ -295,10 +297,10 @@ class MysqlModel implements SqlInterface
         $params = array();
 
         $field_str = $this->parse_fields($fields);
-		$where_str = $this->parse_where($where, $params);
+        $where_str = $this->parse_where($where, $params);
 
         $this->sql = sprintf($one_sql, $field_str, $where_str);
-        $result = $this->prepare( $this->sql )->exec($params)->stmt_fetch();
+        $result = $this->prepare($this->sql)->exec($params)->stmt_fetch();
 
         return $result;
     }
@@ -307,7 +309,7 @@ class MysqlModel implements SqlInterface
      * prepare插入
      *
      * @param string $table 表名称
-     * @param string $data 要处理的数据,跟表中的字段对应
+     * @param array $data 要处理的数据,跟表中的字段对应
      * @param bool $multi <pre>
      * @param array $insert_data
      *  是否批量插入数据,如果是
@@ -321,16 +323,15 @@ class MysqlModel implements SqlInterface
      * @return array|bool
      * @throws FrontException
      */
-    function prepare_insert($table, $data, $multi = false, & $insert_data )
+    function prepare_insert($table, $data, $multi = false, & $insert_data)
     {
         $insert_sql = "INSERT {$table} (%s) VALUE (%s)";
         $field = $value = '';
 
-        if(true === $multi )
-        {
-            $inc_name = $this->get_auto_increment_name( $table );
+        if (true === $multi) {
+            $inc_name = $this->get_auto_increment_name($table);
 
-            if( empty($data ['fields']) || empty($data ['values']) ) {
+            if (empty($data ['fields']) || empty($data ['values'])) {
                 throw new FrontException("data format error!");
             }
 
@@ -341,16 +342,12 @@ class MysqlModel implements SqlInterface
 
             $this->sql = sprintf($insert_sql, rtrim($field, ","), rtrim($value, ","));
             $stmt = $this->prepare($this->sql);
-            $add_data_info = array();
 
-            foreach($data ['values'] as $data_array)
-            {
-                if( $stmt->exec($data_array) )
-                {
+            foreach ($data ['values'] as $data_array) {
+                if ($stmt->exec($data_array)) {
                     $add_data_info = array_combine($data['fields'], $data_array);
-                    if($inc_name)
-                    {
-                        $add_data_info[ $inc_name ] = $this->insert_id();
+                    if ($inc_name) {
+                        $add_data_info[$inc_name] = $this->insert_id();
                     }
 
                     $insert_data[] = $add_data_info;
@@ -358,11 +355,8 @@ class MysqlModel implements SqlInterface
             }
 
             return true;
-        }
-        else
-        {
-            foreach($data as $_field => $_value)
-            {
+        } else {
+            foreach ($data as $_field => $_value) {
                 $field .= "{$_field},";
                 $value .= "?,";
                 $r[] = $_value;
@@ -372,8 +366,7 @@ class MysqlModel implements SqlInterface
             $stmt = $this->prepare($this->sql);
             $id = $stmt->exec($r)->insert_id();
 
-            if($id)
-            {
+            if ($id) {
                 return $id;
             }
 
@@ -389,10 +382,10 @@ class MysqlModel implements SqlInterface
      * @param string $where 查询条件
      * @param int $order 排序
      * @param array $page 分页参数 默认返回50条记录
-     * @param string $group_by
+     * @param int|string $group_by
      * @return array
      */
-    function prepare_find($table, $fields, $where = null, $order = 1, & $page = array('p'=>1, 'limit'=>50), $group_by = 1)
+    function prepare_find($table, $fields, $where = null, $order = 1, & $page = array('p' => 1, 'limit' => 50), $group_by = 1)
     {
         $params = array();
 
@@ -403,21 +396,23 @@ class MysqlModel implements SqlInterface
 
         $total = $this->prepare_getone($table, 'COUNT(*) as total', $where);
 
-        $page['result_count'] = (int) $total ['total'];
-        $page['limit'] = max(1, (int) $page['limit']);
+        $page['result_count'] = (int)$total ['total'];
+        $page['limit'] = max(1, (int)$page['limit']);
         $page['total_page'] = ceil($page['result_count'] / $page['limit']);
-        $page['p'] = max(1, min( $page['p'], $page['total_page']));
+        $page['p'] = max(1, min($page['p'], $page['total_page']));
         $p = ($page['p'] - 1) * $page['limit'];
 
         if (1 !== $group_by) {
             $all_sql = "SELECT %s FROM {$table} WHERE %s GROUP BY %s ORDER BY %s LIMIT %s";
             $this->sql = sprintf($all_sql, $field_str, $where_str, $group_str, $order_str, "{$p}, {$page['limit']}");
-        } else {
+        }
+        else {
             $all_sql = "SELECT %s FROM {$table} WHERE %s ORDER BY %s LIMIT %s";
             $this->sql = sprintf($all_sql, $field_str, $where_str, $order_str, "{$p}, {$page['limit']}");
         }
 
         $result = $this->prepare($this->sql)->exec($params)->stmt_fetch(true);
+
         return $result;
     }
 
@@ -431,7 +426,7 @@ class MysqlModel implements SqlInterface
      * @param int $group_by
      * @return array
      */
-    function prepare_get_all($table, $fields, $where = null, $order = 1, $group_by=1)
+    function prepare_get_all($table, $fields, $where = null, $order = 1, $group_by = 1)
     {
         $params = array();
         $field_str = $this->parse_fields($fields);
@@ -439,26 +434,26 @@ class MysqlModel implements SqlInterface
         $order_str = $this->parse_order($order);
         $group_str = $this->parse_group($group_by);
 
-        if (1 !== $group_by)
-        {
+        if (1 !== $group_by) {
             $all_sql = "SELECT %s FROM {$table} WHERE %s GROUP BY %s ORDER BY %s";
             $this->sql = sprintf($all_sql, $field_str, $where_str, $group_str, $order_str);
         }
-        else
-        {
+        else {
             $all_sql = "SELECT %s FROM {$table} WHERE %s ORDER BY %s";
             $this->sql = sprintf($all_sql, $field_str, $where_str, $order_str);
         }
 
-        $result = $this->prepare($this->sql)->exec( $params )->stmt_fetch(true);
+        $result = $this->prepare($this->sql)->exec($params)->stmt_fetch(true);
+
         return $result;
     }
 
     /**
      * prepare 方式更新
-     * @param $table
-     * @param $data
-     * @param $where
+     *
+     * @param string $table
+     * @param array $data
+     * @param array|string $where
      * @return $this|array|string
      */
     function prepare_update($table, $data, $where)
@@ -466,18 +461,15 @@ class MysqlModel implements SqlInterface
         $up_sql = "UPDATE {$table} SET %s WHERE %s";
         $params = array();
 
-        if(! empty($data) )
-        {
+        if (!empty($data)) {
             $field = '';
-            if(is_array($data))
-            {
-                foreach($data as $d_key => $d_value) {
+            if (is_array($data)) {
+                foreach ($data as $d_key => $d_value) {
                     $field .= "{$d_key} = ? ,";
                     $params [] = strval($d_value);
                 }
             }
-            else
-            {
+            else {
                 $field = $data;
             }
         }
@@ -485,21 +477,21 @@ class MysqlModel implements SqlInterface
         $where_params = array();
         $where_str = $this->parse_where($where, $where_params);
 
-        foreach($where_params as $wp)
-        {
-        	$params [] = $wp;
+        foreach ($where_params as $wp) {
+            $params [] = $wp;
         }
 
         $this->sql = sprintf($up_sql, trim($field, ","), $where_str);
         $this->prepare($this->sql)->exec($params);
+
         return true;
     }
 
     /**
      * 删除数据
      *
-     * @param $table
-     * @param $where
+     * @param string $table
+     * @param string|array $where
      * @param bool $multi 是否批量删除数据
      *      $where = array(
      *          'fields' = array(字段1,字段2,...),
@@ -514,15 +506,13 @@ class MysqlModel implements SqlInterface
     {
         $del_sql = "DELETE FROM %s WHERE %s";
 
-        if(true === $multi )
-        {
-            if( empty($where ['fields']) || empty($where ['values']) ) {
+        if (true === $multi) {
+            if (empty($where ['fields']) || empty($where ['values'])) {
                 throw new FrontException("data format error!");
             }
 
             $where_condition = array();
-            foreach ($where ['fields'] as $d)
-            {
+            foreach ($where ['fields'] as $d) {
                 $where_condition[] = "{$d} = ?";
             }
 
@@ -530,20 +520,19 @@ class MysqlModel implements SqlInterface
             $this->sql = sprintf($del_sql, $table, $where_str);
 
             $stmt = $this->prepare($this->sql);
-            foreach($where ['values'] as $data_array)
-            {
+            foreach ($where ['values'] as $data_array) {
                 $stmt->exec($data_array);
             }
 
             return true;
         }
-        else
-        {
+        else {
             $params = array();
             $where_str = $this->parse_where($where, $params);
 
             $this->sql = sprintf($del_sql, $table, $where_str);
-            $this->prepare( $this->sql )->exec( $params );
+            $this->prepare($this->sql)->exec($params);
+
             return true;
         }
     }
@@ -551,42 +540,39 @@ class MysqlModel implements SqlInterface
     /**
      * 解析字段
      *
-     * @param $fields
+     * @param string|array $fields
      * @return string
      */
-    function parse_fields( $fields )
+    function parse_fields($fields)
     {
-    	if(empty($fields)) {
-    		$field_str = '*';
-    	} else {
-    		if(is_array($fields)) {
-    			$field_str = implode(",", $fields);
-    		} else {
-    			$field_str = $fields;
-    		}
-    	}
+        if (empty($fields)) {
+            $field_str = '*';
+        } else {
+            if (is_array($fields)) {
+                $field_str = implode(",", $fields);
+            }
+            else {
+                $field_str = $fields;
+            }
+        }
 
-    	return $field_str;
+        return $field_str;
     }
 
     /**
      * 解析where
      *
-     * @param $where
-     * @param $params
+     * @param string|array $where
+     * @param array $params
      * @return string
      */
     function parse_where($where, & $params)
     {
-        if(! empty($where))
-        {
-            if( is_array($where) )
-            {
-                foreach($where as $w_key => $w_value)
-                {
+        if (!empty($where)) {
+            if (is_array($where)) {
+                foreach ($where as $w_key => $w_value) {
                     $operator = '=';
-                    if(is_array($w_value))
-                    {
+                    if (is_array($w_value)) {
                         list($operator, $n_value) = $w_value;
                         $w_value = $n_value;
                     }
@@ -595,14 +581,10 @@ class MysqlModel implements SqlInterface
                     $params [] = $w_value;
                 }
                 $where_str = implode(" AND ", $where_str);
-            }
-            else
-            {
+            } else {
                 $where_str = $where;
             }
-        }
-        else
-        {
+        } else {
             $where_str = '1=1';
         }
 
@@ -612,21 +594,20 @@ class MysqlModel implements SqlInterface
     /**
      * 解析order
      *
-     * @param $order
+     * @param string $order
      * @return int|string
      */
     function parse_order($order)
     {
-        if(! empty($order))
-        {
-            if(is_array($order)) {
+        if (!empty($order)) {
+            if (is_array($order)) {
                 $order_str = implode(",", $order);
-            } else {
+            }
+            else {
                 $order_str = $order;
             }
         }
-        else
-        {
+        else {
             $order_str = 1;
         }
 
@@ -636,54 +617,49 @@ class MysqlModel implements SqlInterface
     /**
      * 解析group by
      *
-     * @param $group_by
+     * @param string $group_by
      * @return int
      */
     function parse_group($group_by)
     {
-    	if(! empty($group_by))
-    	{
-    		$group_str = $group_by;
-    	}
-    	else
-    	{
-    		$group_str = 1;
-    	}
-    	return $group_str;
+        if (!empty($group_by)) {
+            $group_str = $group_by;
+        } else {
+            $group_str = 1;
+        }
+
+        return $group_str;
     }
 
     /**
      * 获取数据表信息
      *
-     * @param $table_name
+     * @param string $table_name
      * @return array
      * @throws CoreException
      */
-    function get_db_table_info( $table_name )
+    function get_db_table_info($table_name)
     {
-        $sql  = "SHOW COLUMNS FROM {$table_name}";
+        $sql = "SHOW COLUMNS FROM {$table_name}";
         $info = $this->fetchAll($sql);
-        if($info)
-        {
+        if ($info) {
             return $info;
         }
 
-        throw new CoreException( "SHOW {$table_name} COLUMNS error !" );
+        throw new CoreException("SHOW {$table_name} COLUMNS error !");
     }
 
     /**
      * 获取自增字段名
      *
-     * @param $table_name
+     * @param string $table_name
      * @return bool
      */
-    function get_auto_increment_name( $table_name )
+    function get_auto_increment_name($table_name)
     {
-        $table_info = $this->get_db_table_info( $table_name );
-        foreach($table_info as $ti)
-        {
-            if($ti['Extra'] == 'auto_increment')
-            {
+        $table_info = $this->get_db_table_info($table_name);
+        foreach ($table_info as $ti) {
+            if ($ti['Extra'] == 'auto_increment') {
                 return $ti['Field'];
             }
         }
@@ -696,14 +672,16 @@ class MysqlModel implements SqlInterface
      *
      * @return bool
      */
-    public function commit(){
+    public function commit()
+    {
         return $this->pdo->commit();
     }
 
     /**
      * @return bool
      */
-    public function beginTA(){
+    public function beginTA()
+    {
         return $this->pdo->beginTransaction();
     }
 
@@ -712,7 +690,8 @@ class MysqlModel implements SqlInterface
      *
      * @return bool
      */
-    public function rollBack(){
+    public function rollBack()
+    {
         return $this->pdo->rollBack();
     }
 
@@ -721,7 +700,8 @@ class MysqlModel implements SqlInterface
      *
      * @return string
      */
-    function insert_id(){
+    function insert_id()
+    {
         return $this->pdo->lastInsertId();
     }
 }
